FIX pipelines & add pipelines tests (#409)
当前修复完可以通过pipelines的1000个单测用例。todo：slow部分还有几十个，需要单独再提PR

+++ ppdiffusers/ppdiffusers/pipelines/controlnet/pipeline_controlnet_inpaint.py
@@ -1346,7 +1346,7 @@ class StableDiffusionControlNetInpaintPipeline(
 
                 # predict the noise residual
                 if num_channels_unet == 9:
-                    latent_model_input = paddle.concat([latent_model_input, mask, masked_image_latents], axis=1)
+                    latent_model_input = paddle.concat([latent_model_input, mask.cast(latent_model_input.dtype), masked_image_latents.cast(latent_model_input.dtype)], axis=1)
 
                 noise_pred = self.unet(
                     latent_model_input,
